fix: tree generate

+++ src/main/java/cn/nukkit/network/protocol/types/BlockChangeEntry.java
@@ -3,7 +3,7 @@ package cn.nukkit.network.protocol.types;
 import cn.nukkit.math.BlockVector3;
 
 
-public record BlockChangeEntry(BlockVector3 blockPos, int runtimeID, int updateFlags, long messageEntityID, MessageType messageType) {
+public record BlockChangeEntry(BlockVector3 blockPos, long runtimeID, int updateFlags, long messageEntityID, MessageType messageType) {
     public enum MessageType {
         NONE,
         CREATE,
